Prove the tip signed before Push can light up
- Push lit up the moment fresh counts arrived (`items` clears the
  pushed mark) while `isTipSigned` still described the previous
  tip, so pushing an agent's unsigned commits reached the service's
  refusal as an error the button should have prevented.
- `tipSignature` is now three states: unread until the current tip
  has been checked, and Push dims for anything but proven signed,
  with the hover saying the check is running. Item changes that
  could mean new commits mark it unread and re-run the worktree
  facts, so the button follows the tip rather than trusting the
  last answer.
- The click reads the signature once more, and an unsigned tip
  declines in the footer with `rebaseNeed` refreshed so Rebase
  relights to sign; the service's own refusal stays as the backstop
  no user path reaches.
- The push tests move to `PullRequestsModelTests+Push` and the
  worktree facts to the branch extension, keeping both inside the
  length limits.
